+++ androidx/src/main/java/me/panpf/androidx/app/Servicex.java
@@ -16,11 +16,14 @@
 
 package me.panpf.androidx.app;
 
+import android.accessibilityservice.AccessibilityService;
 import android.app.ActivityManager;
 import android.app.Service;
 import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
+import android.provider.Settings;
+import android.text.TextUtils;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
@@ -28,6 +31,7 @@ import androidx.annotation.Nullable;
 import java.util.ArrayList;
 
 import me.panpf.androidx.content.Contextx;
+import me.panpf.javax.lang.Stringx;
 
 public class Servicex {
 
@@ -88,27 +92,35 @@ public class Servicex {
         context.stopService(new Intent(context, serviceClass));
     }
 
-    // todo add isAccessibilityServiceEnabled
-//    public static boolean isAccessibilityServiceEnabled(@NonNull Context context, @NonNull Class<? extends AccessibilityService> service, final @Nullable String packageName) {
-//        try {
-//            int enable = Settings.Secure.getInt(context.getContentResolver(), Settings.Secure.ACCESSIBILITY_ENABLED, 0);
-//            if (enable == 1) {
-//                String services = Settings.Secure.getString(context.getContentResolver(), Settings.Secure.ENABLED_ACCESSIBILITY_SERVICES);
-//                if (!TextUtils.isEmpty(services)) {
-//                    TextUtils.SimpleStringSplitter split = new TextUtils.SimpleStringSplitter(':');
-//                    split.setString(services);
-//                    final String finalPackageName = packageName != null ? packageName : context.getPackageName();
-//                    final String serviceInfo = finalPackageName + "/" + service.getName();
-//                    while (split.hasNext()) {
-//                        if (split.next().equalsIgnoreCase(serviceInfo)) {
-//                            return true;
-//                        }
-//                    }
-//                }
-//            }
-//        } catch (Throwable e) {
-//            e.printStackTrace();
-//        }
-//        return false;
-//    }
+    /**
+     * Return true if the specified AccessibilityService service has been enabled
+     */
+    public static boolean isAccessibilityServiceEnabled(@NonNull Context context, @NonNull Class<? extends AccessibilityService> service, final @Nullable String packageName) {
+        try {
+            String services = Settings.Secure.getString(context.getContentResolver(), Settings.Secure.ENABLED_ACCESSIBILITY_SERVICES);
+            if (!TextUtils.isEmpty(services)) {
+                TextUtils.SimpleStringSplitter split = new TextUtils.SimpleStringSplitter(':');
+                split.setString(services);
+                final String targetServiceInfo = Stringx.orEmpty(packageName) + "/" + service.getName();
+                while (split.hasNext()) {
+                    final String serviceInfo = split.next();
+                    if (packageName != null && serviceInfo.equalsIgnoreCase(targetServiceInfo)) {
+                        return true;
+                    } else if (packageName == null && serviceInfo.endsWith(targetServiceInfo)) {
+                        return true;
+                    }
+                }
+            }
+        } catch (Throwable e) {
+            e.printStackTrace();
+        }
+        return false;
+    }
+
+    /**
+     * Return true if the specified AccessibilityService service has been enabled
+     */
+    public static boolean isAccessibilityServiceEnabled(@NonNull Context context, @NonNull Class<? extends AccessibilityService> service) {
+        return isAccessibilityServiceEnabled(context, service, null);
+    }
 }